+++ app/src/main/java/io/legado/app/web/HttpServer.kt
@@ -91,6 +91,7 @@ class HttpServer(port: Int) : NanoHTTPD(port) {
                         "/getRssSource" -> RssSourceController.getSource(parameters)
                         "/getRssSources" -> RssSourceController.sources
                         "/getReplaceRules" -> ReplaceRuleController.allRules
+                        "/searchBook" -> BookController.search(parameters)
                         else -> null
                     }
                 }